+++ packages/extension-audio/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tiptap/extension-audio
 
+## 3.21.0
+
+### Patch Changes
+
+- @tiptap/core@3.21.0
+
+## 3.20.6
+
+### Patch Changes
+
+- @tiptap/core@3.20.6
+
 ## 3.20.5
 
 ### Patch Changes