+++ packages/extension-invisible-characters/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tiptap/extension-invisible-characters
 
+## 3.21.0
+
+### Patch Changes
+
+- @tiptap/core@3.21.0
+- @tiptap/extension-text-style@3.21.0
+- @tiptap/pm@3.21.0
+
+## 3.20.6
+
+### Patch Changes
+
+- @tiptap/core@3.20.6
+- @tiptap/extension-text-style@3.20.6
+- @tiptap/pm@3.20.6
+
 ## 3.20.5
 
 ### Patch Changes