+++ packages/extension-mention/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 3.21.0
+
+### Patch Changes
+
+- @tiptap/core@3.21.0
+- @tiptap/pm@3.21.0
+- @tiptap/suggestion@3.21.0
+
+## 3.20.6
+
+### Patch Changes
+
+- @tiptap/core@3.20.6
+- @tiptap/pm@3.20.6
+- @tiptap/suggestion@3.20.6
+
 ## 3.20.5
 
 ### Patch Changes